Update vote.cpp
Here is what I have so far for the vote dataset.  I can't get it to run yet, as it gives a segmentation fault when I do.  It breaks in the print_dataset() method because data.size() returns zero.  If you have any advice on this, please let me know.  I'll continue to try and debug tomorrow.